+++ lib/libc/minimal/Kconfig
@@ -79,6 +79,7 @@ config MINIMAL_LIBC_RAND
 config MINIMAL_LIBC_TIME
 	bool "Time functions"
 	select COMMON_LIBC_TIME if POSIX_CLOCK
+	select COMMON_LIBC_GMTIME_R
 	default y
 	help
 	  Enable time() and gmtime_r() for the minimal libc.
